third_party/blink/web_tests/external/wpt/fetch/cors-rfc1918/non-secure-context.window-expected.txt

This is a testharness.js-based test.
PASS Local non secure page fetches local page.
FAIL Public non secure page fetches local page. assert_equals: expected "failure" but got "success"
Harness: the test ran to completion.
